requirements.txt
cx_oracle
django-admin-interface